+++ buildpacks/nodejs-function-invoker/lib/utils/download.sh
@@ -1,13 +1,13 @@
 #!/usr/bin/env bash
 
 download_yj() {
-	bp_dir=$1
+	app_dir="$1"
 
-	mkdir -p "${bp_dir}/downloads/bin"
+	mkdir -p "${app_dir}/.salesforce_functions/downloads/bin"
 
-	if [[ ! -f "${bp_dir}/downloads/bin/yj" ]]; then
-		curl -Ls https://github.com/sclevine/yj/releases/download/v2.0/yj-linux >"${bp_dir}/downloads/bin/yj" &&
-			chmod +x "${bp_dir}/downloads/bin/yj"
+	if [[ ! -f "${app_dir}/.salesforce_functions/downloads/bin/yj" ]]; then
+		curl -Ls https://github.com/sclevine/yj/releases/download/v2.0/yj-linux >"${app_dir}/.salesforce_functions/downloads/bin/yj" &&
+			chmod +x "${app_dir}/.salesforce_functions/downloads/bin/yj"
 	fi
 }